# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IA Games Ultimate - Data Labeling Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /responses:
    get:
      operationId: listResponses
      tags: [responses]
      summary: List task responses
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: taskId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of responses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
    post:
      operationId: submitResponse
      tags: [responses]
      summary: Submit a task response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResponseBody"
      responses:
        "201":
          description: Response submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResponseResult"
  /responses/{id}:
    get:
      operationId: getResponse
      tags: [responses]
      summary: Get a response by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Response found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          description: Not found

  # ── Datasets ─────────────────────────────────────────
  /datasets:
    get:
      operationId: listDatasets
      tags: [datasets]
      summary: List datasets
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: accessType
          in: query
          schema:
            type: string
            enum: [free, ads, premium]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of datasets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dataset"
    post:
      operationId: createDataset
      tags: [datasets]
      summary: Create a new dataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDatasetBody"
      responses:
        "201":
          description: Dataset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/{id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get a dataset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Dataset found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
        "404":
          description: Not found
    patch:
      operationId: updateDataset
      tags: [datasets]
      summary: Update dataset metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDatasetBody"
      responses:
        "200":
          description: Dataset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
  /datasets/{id}/download:
    post:
      operationId: downloadDataset
      tags: [datasets]
      summary: Record a dataset download (increments download count)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DownloadDatasetBody"
      responses:
        "200":
          description: Download recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataset"
...
  /ads/watch:
    post:
      operationId: watchAd
      tags: [ads]
      summary: Record an ad view and reward the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchAdBody"
      responses:
        "200":
          description: Ad recorded, reward granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchAdResult"
  /ads/tracking/{userId}:
    get:
      operationId: getAdTracking
      tags: [ads]
      summary: Get ad tracking data for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdTracking"

  # ── Rewards ──────────────────────────────────────────
...
components:
  schemas:
...
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        taskId:
          type: integer
        answer:
          type: string
        isCorrect:
          type: ["boolean", "null"]
        responseTimeMs:
          type: integer
        pointsEarned:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, userId, taskId, answer, responseTimeMs, pointsEarned, createdAt]
    SubmitResponseBody:
      type: object
      properties:
        userId:
          type: integer
        taskId:
          type: integer
        answer:
          type: string
        responseTimeMs:
          type: integer
      required: [userId, taskId, answer, responseTimeMs]
    SubmitResponseResult:
      type: object
      properties:
        response:
          $ref: "#/components/schemas/TaskResponse"
        pointsEarned:
          type: integer
        xpEarned:
          type: integer
        accuracyBonus:
          type: boolean
        newLevel:
          type: ["string", "null"]
        streakBonus:
          type: boolean
      required: [response, pointsEarned, xpEarned, accuracyBonus, streakBonus]

    # ── Datasets ──────────────────────────────────────
    Dataset:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        qualityScore:
          type: number
        accessType:
          type: string
          enum: [free, ads, premium]
        price:
          type: ["number", "null"]
        adsRequired:
          type: ["integer", "null"]
        downloadCount:
          type: integer
        size:
          type: ["string", "null"]
        recordCount:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, category, qualityScore, accessType, downloadCount, tags, createdAt]
    CreateDatasetBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        qualityScore:
          type: number
          default: 0
        accessType:
          type: string
          enum: [free, ads, premium]
        price:
          type: ["number", "null"]
        adsRequired:
          type: ["integer", "null"]
        size:
          type: ["string", "null"]
        recordCount:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
      required: [name, description, category, accessType]
    UpdateDatasetBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        qualityScore:
          type: number
        accessType:
          type: string
          enum: [free, ads, premium]
        price:
          type: ["number", "null"]
        adsRequired:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
    DownloadDatasetBody:
      type: object
      properties:
        userId:
          type: ["integer", "null"]
        paymentMethod:
          type: string
          enum: [free, ads, stripe]
      required: [paymentMethod]
...
    WatchAdBody:
      type: object
      properties:
        userId:
          type: integer
        adType:
          type: string
          enum: [rewarded, unlock]
        datasetId:
          type: ["integer", "null"]
      required: [userId, adType]
    WatchAdResult:
      type: object
      properties:
        success:
          type: boolean
        energyRestored:
          type: integer
        pointsEarned:
          type: integer
        datasetUnlocked:
          type: boolean
        adsWatchedToday:
          type: integer
        dailyCapReached:
          type: boolean
      required: [success, energyRestored, pointsEarned, datasetUnlocked, adsWatchedToday, dailyCapReached]
    AdTracking:
      type: object
      properties:
        userId:
          type: integer
        adsWatchedToday:
          type: integer
        totalAdsWatched:
          type: integer
        lastViewTime:
          type: ["string", "null"]
          format: date-time
        dailyCap:
          type: integer
        dailyCapReached:
          type: boolean
      required: [userId, adsWatchedToday, totalAdsWatched, dailyCap, dailyCapReached]

    # ── Rewards ──────────────────────────────────────
...